Rename rust2 -> rust and crate actors2 -> actors
There is no separate idiomatic Rust port, so the "2" was vestigial and
misleading. Rename the directory actors/rust2 -> actors/rust and the crate
actors2 -> actors, and update every reference across code and docs
(examples, tests, generated code, the generator, the schema header, the
crate docs, and the repo-level README/FILE_STRUCTURE).

Verified after the rename:
- cargo build --all-targets --features interop: clean.
- cargo test (feature off) and cargo test --features interop: all pass.
- Examples run: matching_engine (15 tests + demo + bench 4.17 M/s),
  ping_pong (1M round-trips), bench_fast_send (~20 ns/call), bench_pool
  (5.16x), interop codegen round-trip.
- Code generation is byte-deterministic under the new paths.
