+++ actors/rust/DEVELOPER_GUIDE.md
@@ -1,6 +1,6 @@
-# actors2 — Developer Guide
+# actors — Developer Guide
 
-How to build actors with `actors2`. This documents the **actual `rust2` API** (which differs
+How to build actors with `actors`. This documents the **actual `rust` API** (which differs
 from the older channel-based Rust actor crate). See [`README.md`](README.md) for the big picture.
 
 ## 1. Define messages (with integer IDs)
@@ -9,7 +9,7 @@ Every message carries a **compile-time integer id** used for O(1) dispatch. IDs 
 `< 1024`; ids `< 16` are reserved by the framework (`Start = 4`, `Shutdown = 5`).
 
 ```rust
-use actors2::define_message;
+use actors::define_message;
 
 struct Ping { count: i32 }
 define_message!(Ping, 10);
@@ -22,7 +22,7 @@ For hot-path messages sent **async** at high rates, make them **pooled** so deli
 fixed-size blocks instead of calling `malloc` (the third arg is the per-thread cache size):
 
 ```rust
-use actors2::define_pooled_message;
+use actors::define_pooled_message;
 
 struct MarketData { /* ... */ }
 define_pooled_message!(MarketData, 20, 128);
@@ -36,7 +36,7 @@ An actor is a plain struct. Write one handler method per message it accepts, the
 with `handle_messages!`. Handlers take `&mut self`, the typed message, and an `ActorContext`.
 
 ```rust
-use actors2::{handle_messages, ActorContext, ActorRef, Start};
+use actors::{handle_messages, ActorContext, ActorRef, Start};
 
 struct PongActor;
 
@@ -63,10 +63,10 @@ ids are silently ignored (so handling `Start` is optional).
 target.send(Box::new(Ping { count: 1 }), ctx.self_ref());
 
 // synchronous, on the stack: run the target's handler inline on THIS thread; get the reply.
-let reply: Option<actors2::MsgBox> = target.fast_send(&Ping { count: 1 }, ctx.self_ref());
+let reply: Option<actors::MsgBox> = target.fast_send(&Ping { count: 1 }, ctx.self_ref());
 
 // pooled async send (recommended for hot-path messages):
-target.send(actors2::MsgBox::pooled(MarketData { /* ... */ }), ctx.self_ref());
+target.send(actors::MsgBox::pooled(MarketData { /* ... */ }), ctx.self_ref());
 ```
 
 - `send(msg, sender)` — `msg` is `impl Into<MsgBox>`, so `Box::new(x)` (heap) or
@@ -92,7 +92,7 @@ The `Manager` gives each actor its own OS thread (with optional Linux CPU pinnin
 via `ThreadConfig`).
 
 ```rust
-use actors2::{Manager, ThreadConfig};
+use actors::{Manager, ThreadConfig};
 
 fn main() {
     let mut mgr = Manager::new();
@@ -123,7 +123,7 @@ adapter; see [`MATCHING_ENGINE.md`](MATCHING_ENGINE.md)).
 
 ## 7. Cross-language interop (optional, `--features interop`)
 
-`actors2` actors can talk to **C++ Kaspar actors in the same process**, location-transparently —
+`actors` actors can talk to **C++ Kaspar actors in the same process**, location-transparently —
 you call `send` / `fast_send` and never learn the peer is C++. It is off by default; build with
 `--features interop`.
 
@@ -143,7 +143,7 @@ Wire it up at startup: tell the inbound path how to resolve local names, then in
 generated dispatch. After that, a C++ actor is just another `ActorRef`:
 
 ```rust
-use actors2::interop::{generated, register_local_lookup};
+use actors::interop::{generated, register_local_lookup};
 
 let handle = mgr.get_handle();
 register_local_lookup(move |name| handle.get_ref_local(name)); // name -> local ActorRef
@@ -163,6 +163,6 @@ hybrid build are in [`interop/README.md`](interop/README.md).
 
 - Message ids: unique per type, `< 1024`, `< 16` reserved. A duplicate id routes to the wrong
   handler (debug builds assert; release drops the message).
-- `actors2` is **in-process only** — no remote/ZMQ transport or actor registry. The one
+- `actors` is **in-process only** — no remote/ZMQ transport or actor registry. The one
   cross-boundary path is same-process C++ interop (above), behind the `interop` feature.
 - The mailbox is blocking (`Mutex`+`Condvar`); the fast path is `fast_send` (no queue).
