+++ actors/rust/src/interop/generated.rs
